C:\Users\zainu\Desktop\курс\1proekt\1proekt\Ozhaev Zaynudin cartoon.cpp -o C:\Users\zainu\Desktop\курс\1proekt\1proekt\Ozhaev Zaynudin cartoon.exe  -Wshadow -Winit-self -Wredundant-decls -Wcast-align -Wundef -Wfloat-equal -Winline -Wunreachable-code -Wmissing-declarations -Wmissing-include-dirs -Wswitch-enum -Wswitch-default -Weffc++ -Wmain -Wextra -Wall -g -pipe -fexceptions -Wcast-qual -Wconversion -Wctor-dtor-privacy -Wempty-body -Wformat-security -Wformat=2 -Wignored-qualifiers -Wlogical-op -Wmissing-field-initializers -Wnon-virtual-dtor -Woverloaded-virtual -Wpointer-arith -Wsign-promo -Wstack-usage=8192 -Wstrict-aliasing -Wstrict-null-sentinel -Wtype-limits -Wwrite-strings -D_DEBUG -D_EJUDGE_CLIENT_SIDE -D_EJC

+++ Ozhaev Zaynudin cartoon.cpp
@@ -3,7 +3,7 @@
 void DrawMir();
 void DrawHouse2 (int x, int y);
 void DrawHouse3 (int x, int y);
-void DrawTree1  (int x, int y, long sizeX, double sizeY, COLORREF TreeColor);
+void DrawTree1  (int x, int y, int sizeX, double sizeY, COLORREF TreeColor);
 void DrawTree2  (int x, int y, double sizeX, double sizeY, COLORREF TreeColor);
 void DrawSun    (int x, double y, double sizeX, double sizeY, COLORREF SunColor);
 void DrawGirl   (int x, int y, double sizeX, double sizeY, COLORREF GirlColor,
